Fixed handling of FTP/HTTP path when using the "@" Added border effect for pictures (next step: rounded corners)
Added stub for new feature: It should be possible to hold shift (or something else) while making a screenshot to give the user the one time option to get something else (e.g. standard is upload image and this time he/she wants it copied to clipboard. or this time he wants to use the image editor)