addressed the following issues thanks to tux0r
- The "index out of bounds" error I had before is gone.
- Trying to log in when already logged in is possible (?) but leads to a
"parameter must not be NULL" error message. You might want to disable
the login area then.
- Uploading a screenshot automatically to a predefined folder seems to
work as well as creating folders.

HOWEVER:

Choosing a folder and clicking "Remove folder" STILL duplicates that
folder in the drop-down list - resulting in the already known "index out
of bounds" error when choosing the duplicate accidentally.